🧪 Steward: Harden accessibility assertions and relocate database-dependent unit tests
- Replace brittle `assertSeeHtml` with `assertSeeInOrder` in `AdminListAccessibilityTest` to avoid coupling with Tailwind classes.
- Relocate `MeetingPhotoMigrationServiceTest` and `MeetingShowPresenterTest` from `Unit/` to `Integration/` as they hit the database.
- Consolidate redundant migration tests and ensure parallel test isolation using dynamic slugs.
